+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - Commands `pcs resource|stonith|booth delete|remove` and
   `pcs cluster node delete-remote|remove-remote|delete-guest|remove-guest`
   now warn about omitting live cluster actions when '-f' is used ([RHEL-76157])
-- Add lib command `cib.set_description` to API v1 and v2 ([RHEL-7670])
+- Add lib commands `cib.get_description` and `cib.set_description` to API v1 and
+  v2 ([RHEL-7670])
 
 ### Fixed
 - Commands `pcs acl role create` and `pcs acl permission add` now print